Remove fixed nav options, and simplify component
As of the requests on the Trello card below, the fixed display options are no longer needed on the `kano-nav`, so the fixed properties, styles, and all related functionality has been removed. Other optional aspects and the `--kano-nav` mixin have been removed to simplify things in line with current needs. In cases where the nav is to be fixed, this can be styled from the container app.

Trello card: https://trello.com/c/hcNI7mIL/6-1-make-the-primary-nav-visible-on-scroll-on-kano-me