Add a success property
This defaults to true, so should not affect behaviour in any apps not using it. This allows the parent app to control flag up whether the authentication has been successful though, and disable the `Done` button (with spinner) until flagged as successful by the parent app.

This sould help to fix the following bug:
https://trello.com/c/Z7sfeWFv/54-kano-world-onboarding-creating-a-user-after-onboarding-does-not-log-him-in-automatically