Update tracking behavior to use localStorage
Since Kano Code can access localStorage, there's no need for schema and sessionId to be passed in URL params. But Kano Code should not clear the saved sessionId, as this will be needed by the Electron app to preserve the previous session. Kano Code will now have a `preserveSavedSession` property to prevent this being cleared.