Abort share fetches if the current user has changed
Add a `resetRequired` flag to override the current fetch request and reset if the current user has changed while the request is being made. If the component is not fetching, and has already been populated, reset as normal. This feels slightly clumsy, but corrects an issue where users were being shown the shares of other users, or concatenations of user shares because the `userId` was only updating when the fetch request was already in progress. Setting a `resetRequired` flag allows the results of the existing request to be ignored, and a new request made with the updated user details.

Should fix this bug: https://trello.com/c/5v4YpJjr/452-first-time-going-in-on-a-user-who-does-not-have-any-pixel-kit-shares-you-see-other-peoples-pixel-kit-shares