Force tombstoned share cards to have default cursors and hidden usernames
When `kano-share-card`s have the `tombstone` property, nothing should be clickable, so the cursor should be `default`. The username should also not appear when hovering over the `#user` div, as this will be blank when tombstoned. `!important` seemed appropriate here to override the attributes on all children, and avoid unnecessary specificity.

Fixes this bug: https://trello.com/c/XxsnuB1U/484-hovering-over-user-icon-brings-up-a-blank-popup