Set feed to complete if there are no more pages to fetch
Previously, we were only setting the feed as complete when calling `_loadMoreData` again. But this would not be called if only a few entries had loaded and the treshold was never getting triggered. Also, we should not have been implying that there were more shares to load if this wasn't the case. This sets the feed as complete if there is no next page.