Check progress object in the user profile for unlocked badges
For each category in the `user.profile.progress` Object, check whether that category is a badge category. If so, iterate over all the badges in that category and push unlocked badges into the `unlockedBadges` array. Use the length of this array to set the number of badges on the user stats. Also corrects the `following` key, which was previously – erroneously – `followers`, so that this displays the correct values.

Trello card: https://trello.com/c/yW6nfafJ/636-1-hook-up-profile-badge-count-to-new-medals