Introduce promise-based method to set up the session or reject with no token
The previous strategy when checking for the user when showing the device page was causing issues with the promise resolution and the actual data being checked being out of sync. To try to avoid any problems, I have introduced a new promise based method separate from other methods to ensure that the user has been initialized. The also avoids errors in Polymer for Promise rejections not being handled when promise functions are used as observers.

This should fix the issues in the Trello card below, where the user was still being checked before being initialized and the user was being prompted to log in unnecessarily.

Trello card: https://trello.com/c/CHFFCnjC/1004-p2-rc24-win10-rpk-about-every-3-4-times-you-start-kca-you-have-to-log-in-again